pyproject.toml
[build-system]
requires = ["hatchling >= 1.26"]
build-backend = "hatchling.build"

[project]
name = "spots_cli"
version = "0.1.0"
authors = [
  { name="Marlon Baatjes", email="marlonbaatjes62@gmail.com" }
]
description = "Spots is a modular music coordination library that unifies multiple providers (e.g. Deezer, Spotify, YouTube) into a single, consistent interface."
readme = "README.md"
requires-python = ">=3.9"
classifiers = [
  "Development Status :: 5 - Production/Stable",
  "Environment :: Console",
  "Intended Audience :: End Users/Desktop",
  "Operating System :: OS Independent",
  "Programming Language :: Python :: 3.14"
]
license = "GPL-3.0-or-later"
dependencies = [
  "beautifulsoup4",
  "python-dotenv",
  "google-api-python-client",
  "google-auth-oauthlib",
  "lyricsgenius",
  "moviepy",
  "mutagen",
  "requests",
  "spotipy",
  "tenacity",
  "yt-dlp",
  "typer"
]

[project.urls]
Homepage = "https://github.com/mdesignscode/spots_flask"
Issues = "https://github.com/mdesignscode/spots_flask"

[tool.setuptools.packages.find]
where = ["src"]

[tool.mypy]
explicit_package_bases = true

[project.scripts]
spots_cli = "spots_cli.cli:main"

[tool.setuptools.package-data]
spots_cli = ["spot_cli.config/client_secrets.json"]
